// api/generated/model/exerciseScoresTrainingFormDtoUnit.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * LgymApi.Api
 * OpenAPI spec version: 1.0
 */

export type ExerciseScoresTrainingFormDtoUnit = typeof ExerciseScoresTrainingFormDtoUnit[keyof typeof ExerciseScoresTrainingFormDtoUnit];


export const ExerciseScoresTrainingFormDtoUnit = {
  Unknown: 'Unknown',
  Kilograms: 'Kilograms',
  Pounds: 'Pounds',
} as const;